+++ Sources/DocumentKit/_Deprecated/DocumentGroup+Deprecated.swift
@@ -11,3 +11,8 @@ public extension DocumentGroup {
         tryCustomizeDocumentBrowser(delay: delay, action)
     }
 }
+
+public extension DocumentGroupModal {
+
+    
+}